feat: Add support for a second footer logo
Introduces a new `footerLogo2` field to the `LeagueData` type and integrates it into the `App` component. This allows users to specify an additional logo for the footer section of the generated standings image. The `handleLeagueDataChange` function has also been updated to accept `null` values for string fields to accommodate optional logos.